+++ sklearn/tests/test_pipeline.py
@@ -12,7 +12,7 @@ from sklearn.utils.testing import assert_array_equal
 from sklearn.utils.testing import assert_array_almost_equal
 
 from sklearn.base import BaseEstimator, clone
-from sklearn.pipeline import Pipeline, FeatureUnion
+from sklearn.pipeline import Pipeline, FeatureUnion, make_pipeline, make_union
 from sklearn.svm import SVC
 from sklearn.linear_model import LogisticRegression
 from sklearn.feature_selection import SelectKBest, f_classif
@@ -230,6 +230,15 @@ def test_feature_union():
     assert_equal(X_transformed.shape, (X.shape[0], 8))
 
 
+def test_make_union():
+    pca = PCA()
+    mock = TransfT()
+    fu = make_union(pca, mock)
+    names, transformers = zip(*fu.transformer_list)
+    assert_equal(names, ("pca", "transft"))
+    assert_equal(transformers, (pca, mock))
+
+
 def test_pipeline_transform():
     # Test whether pipeline works with a transformer at the end.
     # Also test pipeline.transform and pipeline.inverse_transform
@@ -262,6 +271,22 @@ def test_pipeline_fit_transform():
     X_trans = pipeline.fit_transform(X, y)
     X_trans2 = transft.fit(X, y).transform(X)
     assert_array_almost_equal(X_trans, X_trans2)
+
+
+def test_make_pipeline():
+    t1 = TransfT()
+    t2 = TransfT()
+
+    pipe = make_pipeline(t1, t2)
+    assert_true(isinstance(pipe, Pipeline))
+    assert_equal(pipe.steps[0][0], "transft-1")
+    assert_equal(pipe.steps[1][0], "transft-2")
+
+    pipe = make_pipeline(t1, t2, FitParamT())
+    assert_true(isinstance(pipe, Pipeline))
+    assert_equal(pipe.steps[0][0], "transft-1")
+    assert_equal(pipe.steps[1][0], "transft-2")
+    assert_equal(pipe.steps[2][0], "fitparamt")
 
 
 def test_feature_union_weights():